Add PDF slides for “What’s new in Laravel 5?”
Slides of the talk that has been given at the Laravel Brussels Meetup group of December 2014 at Pianofabriek.

These slides were created more as a support for what was said orally rather than as an exhaustive copy of the talk’s contents. As such, they may be less useful for people who did not attend since several things that we covered are not mentioned in the slides.